Refactor components in schedule list
